io: add paging print_zettel function
Impliment print_zettel(). Takes arg for where config will go,
eventually, as well as the un-canonicalized pathname of the zettel.

print_zettel() uses a fork()/wait() model to call the specified
pager (the default is bat(1) for syntax highlighting, though this may
change at a later date) and falls back to a simple highly optimized
chunk-based read and print function if bat(1) is not found.