conf: start writing config system
One of the greatest needs of a flexible application is a configuration
method. ze(1) will have support for a simple

key=value

type config file. The location is yet to be determined.

The config system is essentially a singly-linked list of key-value
pairs as found inside of the config file. This means that each
retreival and addition can be done in O(n) time, with the addition of
calls to malloc().