build: add MacOS lmdb file
Most Linux distributions provide a pkg-config file with libraries when
they are installed. However, this practice is not normal on
MacOS. This pkg-config file is hand-written and not guarenteed to be
accurate; it assumes a homebrew install of lmdb. Change the version
number if yours breaks; this happens to be the version on this
developer's mac.

If, while building on a mac, the configure script breaks and says that
``LMDB cannot be found'', set the environment variable PKG_CONFIG_PATH
to ".." (or wherever the lmdb.pc file is relative to your build) and
run the configure script again. If it still breaks, make sure that the
version number in $prefix is correct. If it isn't, then change it and
run the configure script a third time.