tests: improve documentation
Add documentation to test functions and establish naming pattern for
future tests.

Tests have to be cross-platform and portable. However, many of the
functions in ``ze'' use files and similar structures to inform their
behavior. Therefore, for any file that is used in a test that is
passed to home_prefix(), the name should start with at least six
characters of garbage (for example, "/ZZZZZZtest.id" in
test_get_id_no_file()) to prevent stomping on user's files. If the
file is not being passed to home_prefix(), use the standard library
function tmpfile() in <stdio.h>