conf: fix null dereference bug
In parse_config(), if a file was found but empty, it would not
correctly set cfg->head to null, resulting in a null pointer
dereference when get_config() is called. This has been fixed by moving
setting cfg->head to NULL before the stat call; it ``rolls over'' into
the rest of the function if it is not overwritten by valid functions.