conf: possible new parser
Experiment with a yacc/lex (or more realistically, bison/flex) based
parser for config. This would allow for a more flexible and powerful
config grammar, at the cost of a slightly more complicated build
system. I intend to tinker with the parser and lexer more, to make
sure there aren't any regressions, and that it builds on all the
platforms available to me.